Rename OneTimePasswordValidator to OneTimePasswordHelper and consolidate the duplicated GenerateOneTimePassword logic there
